fix(order): delete method and tests

+++ src/core/application/useCases/pedido/index.ts
@@ -66,6 +66,12 @@ export class PedidoUseCase implements IPedidoUseCase {
     }
 
     public async delete(id: string): Promise<void> {
+        const doesPedidoExists = await this.pedidoRepository.getById(id);
+
+        if (!doesPedidoExists) {
+            throw new Error("Pedido não encontrado");
+        }
+
         this.pedidoRepository.delete(id);
     }
 }
